refactor: library sources directory refactored auto -> lib
