+++ src/app.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AppService } from './app.service';
 import { AppController } from './app.controller';
-import { ComponentScan } from './auto/decorator/component-scan.decorator';
+import { ComponentScan } from './lib/decorator/component-scan.decorator';
 
 @ComponentScan()
 @Module({
