+++ src/lib/auto.module.ts
