+++ src/lib/decorator/component-scan.decorator.ts
