+++ test/fixture.ts
@@ -1,4 +1,4 @@
-import { AutoInjectable } from '../src/auto/decorator/auto.injectable.decorator';
+import { AutoInjectable } from '../src/lib/decorator/auto.injectable.decorator';
 
 @AutoInjectable()
-export class AutoInjectableFixture1 {}
+export class AutoInjectableFixture {}
